+++ examples/3qubit_gates.jl
@@ -5,7 +5,7 @@ function decompose_RX_on_q3()
     params = Dict{String, Any}(
     
     "num_qubits" => 3, 
-    "depth" => 3,    
+    "maximum_depth" => 3,    
 
     "elementary_gates" => ["U3_3", "Identity"], 
     "target_gate" => QCO.kron_single_qubit_gate(3, QCO.RXGate(π/4), "q3"),
@@ -17,7 +17,7 @@ function decompose_RX_on_q3()
     "objective" => "minimize_depth", 
     "decomposition_type" => "exact",
     
-    "optimizer" => "cplex",
+    "optimizer" => "gurobi",
     "optimizer_presolve" => false, #turning this true will give infeasiblity in cplex - most probably a bug in cplex's presolve
     
     )
@@ -33,17 +33,20 @@ function decompose_toffoli()
     params = Dict{String, Any}(
     
     "num_qubits" => 3, 
-    "depth" => 15,    
+    "maximum_depth" => 15, 
 
     "elementary_gates" => ["T_1", "T_2", "T_3", "H_3", "CNot_1_2", "CNot_1_3", "CNot_2_3", "Tdagger_1", "Tdagger_2", "Tdagger_3", "Identity"], 
     "target_gate" => QCO.ToffoliGate(),
     "input_circuit" => toffoli_circuit(),
+
+    "set_cnot_lower_bound" => 6,
+    "set_cnot_upper_bound" => 6,
     
     "objective" => "minimize_depth", 
     "decomposition_type" => "exact",
     "relax_integrality" => false,
     
-    "optimizer" => "cplex",
+    "optimizer" => "gurobi",
     
     )
 
@@ -57,8 +60,8 @@ function decompose_toffoli_using_kronecker()
  
     params = Dict{String, Any}(
     
-    "num_qubits" => 3, 
-    "depth" => 12,    
+    "num_qubits" => 3,
+    "maximum_depth" => 12,
   
     "elementary_gates" => ["T_3", "H_3", "CNot_1_2", "CNot_1_3", "CNot_2_3", "Tdagger_3", "I_1xT_2xT_3", "CNot_1_2xH_3", "T_1xTdagger_2xI_3", "Identity"], 
     "target_gate" => QCO.ToffoliGate(),
@@ -66,8 +69,10 @@ function decompose_toffoli_using_kronecker()
     "objective" => "minimize_depth", 
     "decomposition_type" => "exact",
     "relax_integrality" => false,
+
+    "set_cnot_lower_bound" => 6,
     
-    "optimizer" => "cplex",
+    "optimizer" => "gurobi",
     
     )
 
@@ -83,7 +88,7 @@ function decompose_toffoli_with_controlled_gates()
     params = Dict{String, Any}(
     
     "num_qubits" => 3,
-    "depth" => 5,
+    "maximum_depth" => 5,
 
     "elementary_gates" => ["CV_1_3", "CV_2_3", "CV_1_2", "CVdagger_1_3", "CVdagger_2_3", "CVdagger_1_2", "CNot_2_1", "CNot_1_2", "Identity"],
     # "elementary_gates" => ["CV_1_3", "CV_2_3", "CVdagger_1_3", "CNot_1_2", "CNot_2_1", "Identity"], 
@@ -94,7 +99,7 @@ function decompose_toffoli_with_controlled_gates()
     "decomposition_type" => "exact",
     "relax_integrality" => false,
     
-    "optimizer" => "cplex",
+    "optimizer" => "gurobi",
     )
 
     return params
@@ -125,14 +130,14 @@ function decompose_CNot_1_3()
 
     params = Dict{String, Any}(
     "num_qubits" => 3,
-    "depth" => 8,
+    "maximum_depth" => 8,
 
     # "elementary_gates" => ["CNot_1_2", "CNot_2_3", "Identity"], 
     "elementary_gates" => ["H_1", "H_3", "H_2", "CNot_2_1", "CNot_3_2", "Identity"],
     "target_gate" => QCO.get_full_sized_gate("CNot_1_3", 3),
 
     "objective" => "minimize_depth", 
-    "optimizer" => "cplex"   
+    "optimizer" => "gurobi"   
     )
 
     return params
@@ -142,14 +147,14 @@ function decompose_FredkinGate()
 
     params = Dict{String, Any}(
     "num_qubits" => 3,
-    "depth" => 7,
+    "maximum_depth" => 7,
 
     # Reference: https://doi.org/10.1103/PhysRevA.53.2855
     "elementary_gates" => ["CV_1_2", "CV_2_3", "CV_1_3", "CVdagger_1_2", "CVdagger_2_3", "CVdagger_1_3", "CNot_1_2", "CNot_3_2", "CNot_2_3", "CNot_1_3", "Identity"],
     "target_gate" => QCO.CSwapGate(), #also Fredkin
 
     "objective" => "minimize_depth", 
-    "optimizer" => "cplex"
+    "optimizer" => "gurobi"
     )
     
     return params
@@ -171,13 +176,13 @@ function decompose_toffoli_left()
 
     params = Dict{String, Any}(
     "num_qubits" => 3,
-    "depth" => 7,
+    "maximum_depth" => 7,
 
     "elementary_gates" => ["H_3", "T_1", "T_2", "T_3", "Tdagger_1", "Tdagger_2", "Tdagger_3", "CNot_1_2", "CNot_2_3", "CNot_1_3", "Identity"],
     "target_gate" => target_gate(), 
 
     "objective" => "minimize_depth", 
-    "optimizer" => "cplex"
+    "optimizer" => "gurobi"
     )
     
     return params
@@ -201,13 +206,13 @@ function decompose_toffoli_right()
 
     params = Dict{String, Any}(
     "num_qubits" => 3,
-    "depth" => 8,
+    "maximum_depth" => 8,
 
     "elementary_gates" => ["H_3", "T_1", "T_2", "T_3", "Tdagger_1", "Tdagger_2", "Tdagger_3", "CNot_1_2", "CNot_2_3", "CNot_1_3", "Identity"],
     "target_gate" => target_gate(), 
 
     "objective" => "minimize_depth", 
-    "optimizer" => "cplex"
+    "optimizer" => "gurobi"
     )
     
     return params
@@ -229,13 +234,13 @@ function decompose_miller()
 
     params = Dict{String, Any}(
         "num_qubits" => 3,
-        "depth" => 8,
+        "maximum_depth" => 8,
     
         "elementary_gates" => ["CV_1_3", "CV_2_3", "CVdagger_2_3", "CNot_1_2", "CNot_3_1", "CNot_3_2", "Identity"],
         "target_gate" => target_gate(), 
     
-        "objective" => "minimize_cnot", 
-        "optimizer" => "cplex"
+        "objective" => "minimize_depth", 
+        "optimizer" => "gurobi"
         )
         
         return params