+++ examples/ex1.jl
@@ -36,7 +36,7 @@ end
 
 params = Dict{String, Any}(
 "num_qubits" => 3,
-"depth" => 7,
+"maximum_depth" => 7,
 
 # "elementary_gates" => ["T_1", "T_2", "T_3", "H_3", "CNot_1_2", "CNot_1_3", "CNot_2_3", "Tdagger_2", "Tdagger_3", "Identity"],
 "elementary_gates" => ["CV_1_2", "CV_2_3", "CV_1_3", "CVdagger_1_2", "CVdagger_2_3", "CVdagger_1_3", "CNot_1_2", "CNot_3_2", "CNot_2_3", "CNot_1_3", "Identity"],