+++ examples/run_examples.jl
@@ -1,7 +1,7 @@
 import QuantumCircuitOpt as QCO
 using JuMP
 using CPLEX
-using LinearAlgebra
+using Gurobi
 using Cbc
 
 include("solver.jl")
@@ -15,8 +15,8 @@ decompose_gates = ["decompose_hadamard",
                    "decompose_controlled_V",
                    "decompose_controlled_H",
                    "decompose_controlled_H_with_R",
-                   "decompose_magic_M", 
-                   "decompose_magic_M_using_SHCnot",
+                   "decompose_magic", 
+                   "decompose_magic_using_SHCnot",
                    "decompose_S", 
                    "decompose_revcnot", 
                    "decompose_revcnot_with_U", 
@@ -27,7 +27,7 @@ decompose_gates = ["decompose_hadamard",
                    "decompose_iSwapGate",
                    "decompose_RX_on_q3"]
 
-decompose_gates = ["decompose_double_toffoli"]
+# decompose_gates = ["decompose_toffoli_right"]
 
 #----------------------------------------------#
 #      Quantum Circuit Optimization model      #
@@ -36,11 +36,15 @@ result_qc = Dict{String,Any}()
 
 for gates in decompose_gates 
     params = getfield(Main, Symbol(gates))()
-    # qcm_optimizer = get_solver(params)
-    qcm_optimizer = JuMP.optimizer_with_attributes(CPLEX.Optimizer)
+    
+    qcm_optimizer = get_solver(params)
+    # qcm_optimizer = JuMP.optimizer_with_attributes(Cbc.Optimizer)
+    # qcm_optimizer = JuMP.optimizer_with_attributes(CPLEX.Optimizer)
+    # qcm_optimizer = JuMP.optimizer_with_attributes(Gurobi.Optimizer, "Presolve" => 1)
+    
     global result_qc = QCO.run_QCModel(params, 
                                        qcm_optimizer, 
                                        model_type = "compact_formulation",
-                                       convex_hull_complex_gate_constraints = true)
+                                       convex_hull_gate_constraints = false)
 end