+++ src/constraints.jl
@@ -5,7 +5,7 @@
 function constraint_single_gate_per_depth(qcm::QuantumCircuitModel)
 
     num_gates = size(qcm.data["gates_real"])[3]
-    depth     = qcm.data["depth"]
+    depth     = qcm.data["maximum_depth"]
     
     JuMP.@constraint(qcm.model, [d=1:depth], sum(qcm.variables[:z_onoff_var][n,d] for n=1:num_gates) == 1)
     
@@ -15,7 +15,7 @@ end
 function constraint_gates_onoff_per_depth(qcm::QuantumCircuitModel)
 
     num_gates = size(qcm.data["gates_real"])[3]
-    depth     = qcm.data["depth"]
+    depth     = qcm.data["maximum_depth"]
 
     JuMP.@constraint(qcm.model, [d=1:depth], qcm.variables[:G_var][:,:,d] .== 
                                     sum(qcm.variables[:z_onoff_var][n,d] * qcm.data["gates_real"][:,:,n] for n=1:num_gates))
@@ -36,7 +36,7 @@ end
 function constraint_gate_intermediate_products(qcm::QuantumCircuitModel)
 
     num_gates = size(qcm.data["gates_real"])[3]
-    depth     = qcm.data["depth"]
+    depth     = qcm.data["maximum_depth"]
 
     JuMP.@constraint(qcm.model, [d=1:(depth-1)], qcm.variables[:U_var][:,:,d] .== 
                                 sum(qcm.variables[:V_var][:,:,n,d] * qcm.data["gates_real"][:,:,n] for n=1:num_gates))
@@ -52,7 +52,7 @@ end
 
 function constraint_gate_target_condition(qcm::QuantumCircuitModel)
 
-    depth   = qcm.data["depth"]
+    depth   = qcm.data["maximum_depth"]
     num_gates = size(qcm.data["gates_real"])[3]
     decomposition_type = qcm.data["decomposition_type"]
     
@@ -70,7 +70,7 @@ end
 
 function constraint_complex_to_real_symmetry(qcm::QuantumCircuitModel)
 
-    depth  = qcm.data["depth"]
+    depth  = qcm.data["maximum_depth"]
     n_r    = size(qcm.data["gates_real"])[1]
     n_c    = size(qcm.data["gates_real"])[2]
 
@@ -92,7 +92,7 @@ end
 
 function constraint_gate_product_linearization(qcm::QuantumCircuitModel)
 
-    depth   = qcm.data["depth"]
+    depth   = qcm.data["maximum_depth"]
     n_r     = size(qcm.data["gates_real"])[1]
     n_c     = size(qcm.data["gates_real"])[2]
     num_gates = size(qcm.data["gates_real"])[3]
@@ -134,7 +134,7 @@ end
 function constraint_gate_intermediate_products_compact(qcm::QuantumCircuitModel)
 
     num_gates = size(qcm.data["gates_real"])[3]
-    depth   = qcm.data["depth"]
+    depth   = qcm.data["maximum_depth"]
     
     JuMP.@constraint(qcm.model, [d=2:(depth-1)], qcm.variables[:U_var][:,:,d] .== 
                                         sum((qcm.variables[:zU_var][:,:,n,(d-1)]) * qcm.data["gates_real"][:,:,n] for n=1:num_gates))
@@ -144,7 +144,7 @@ end
 
 function constraint_gate_target_condition_compact(qcm::QuantumCircuitModel)
 
-    depth   = qcm.data["depth"]
+    depth   = qcm.data["maximum_depth"]
     num_gates = size(qcm.data["gates_real"])[3]
     decomposition_type = qcm.data["decomposition_type"]
 
@@ -164,7 +164,7 @@ end
 
 function constraint_complex_to_real_symmetry_compact(qcm::QuantumCircuitModel)
 
-    depth  = qcm.data["depth"]
+    depth  = qcm.data["maximum_depth"]
     n_r    = size(qcm.data["gates_real"])[1]
     n_c    = size(qcm.data["gates_real"])[2]
 
@@ -176,7 +176,7 @@ end
 
 function constraint_commutative_gate_pairs(qcm::QuantumCircuitModel)
     
-    depth  = qcm.data["depth"]
+    depth  = qcm.data["maximum_depth"]
     z_onoff_var  = qcm.variables[:z_onoff_var]
 
     commute_pairs, commute_pairs_prodIdentity = QCO.get_commutative_gate_pairs(qcm.data["gates_dict"])
@@ -206,7 +206,7 @@ end
 function constraint_involutory_gates(qcm::QuantumCircuitModel)
 
     gates_dict  = qcm.data["gates_dict"]
-    depth       = qcm.data["depth"]
+    depth       = qcm.data["maximum_depth"]
     z_onoff_var = qcm.variables[:z_onoff_var]
 
     involutory_gates = QCO.get_involutory_gates(gates_dict)
@@ -226,7 +226,7 @@ end
 function constraint_redundant_gate_product_pairs(qcm::QuantumCircuitModel)
 
     gates_dict  = qcm.data["gates_dict"]
-    depth       = qcm.data["depth"]
+    depth       = qcm.data["maximum_depth"]
     z_onoff_var = qcm.variables[:z_onoff_var]
 
     redundant_pairs = QCO.get_redundant_gate_product_pairs(gates_dict)
@@ -246,7 +246,7 @@ end
 function constraint_idempotent_gates(qcm::QuantumCircuitModel)
 
     gates_dict  = qcm.data["gates_dict"]
-    depth       = qcm.data["depth"]
+    depth       = qcm.data["maximum_depth"]
     z_onoff_var = qcm.variables[:z_onoff_var]
 
     idempotent_gates = QCO.get_idempotent_gates(gates_dict)
@@ -266,7 +266,7 @@ end
 function constraint_identity_gate_symmetry(qcm::QuantumCircuitModel)
 
     gates_dict  = qcm.data["gates_dict"]
-    depth       = qcm.data["depth"]
+    depth       = qcm.data["maximum_depth"]
     z_onoff_var = qcm.variables[:z_onoff_var]
 
     identity_idx = []
@@ -285,6 +285,27 @@ function constraint_identity_gate_symmetry(qcm::QuantumCircuitModel)
     return
 end
 
+function constraint_cnot_gate_bounds(qcm::QuantumCircuitModel)
+
+    cnot_idx    = qcm.data["cnot_idx"]
+    depth       = qcm.data["maximum_depth"]
+    z_onoff_var = qcm.variables[:z_onoff_var]
+
+    if !isempty(cnot_idx)
+        if "cnot_lower_bound" in keys(qcm.data)
+            JuMP.@constraint(qcm.model, sum(z_onoff_var[n,d] for n in cnot_idx, d=1:depth) >= qcm.data["cnot_lower_bound"])
+            Memento.info(_LOGGER, "Applied CNot-gate lower bound constraint")
+        end
+        
+        if "cnot_upper_bound" in keys(qcm.data)
+            JuMP.@constraint(qcm.model, sum(z_onoff_var[n,d] for n in cnot_idx, d=1:depth) <= qcm.data["cnot_upper_bound"])
+            Memento.info(_LOGGER, "Applied CNot-gate upper bound constraint")
+        end
+    end
+
+    return
+end
+
 function constraint_convex_hull_complex_gates(qcm::QuantumCircuitModel)
 
     if !qcm.data["are_gates_real"] 
@@ -297,7 +318,7 @@ function constraint_convex_hull_complex_gates(qcm::QuantumCircuitModel)
         gates_dict = qcm.data["gates_dict"]
 
         num_gates = size(gates_real)[3]
-        depth     = qcm.data["depth"]
+        depth     = qcm.data["maximum_depth"]
         n_r       = size(gates_dict["1"]["matrix"])[1]
         n_c       = size(gates_dict["1"]["matrix"])[2]
 
@@ -315,24 +336,23 @@ function constraint_convex_hull_complex_gates(qcm::QuantumCircuitModel)
 
                     push!(vertices_coord, (re, im))
                 end
-
+                
                 if (isapprox(minimum([x[1] for x in vertices_coord]), maximum([x[1] for x in vertices_coord]), atol = 1E-6)) || (isapprox(minimum([x[2] for x in vertices_coord]), maximum([x[2] for x in vertices_coord]), atol = 1E-6))
                     continue
                 end
 
                 vertices = Vector{Tuple{<:Number, <:Number}}()
-
+                
                 if length(vertices_coord) == 2 
-
+                    
                     for l in vertices_coord
                         push!(vertices, (l[1], l[2]))
                     end
 
                     slope, intercept = QCO._get_constraint_slope_intercept(vertices[1], vertices[2])
-                    
+
                     if !isinf(slope)
                         if isapprox(abs(slope), 0, atol=1E-6)
-
                             JuMP.@constraint(qcm.model, [d=1:depth], 
                                             sum(gates_real[(2*I-1),(2*J), n_g] * z_onoff_var[n_g,d] for n_g = 1:num_gates) - intercept == 0)
                         else
@@ -361,7 +381,7 @@ function constraint_convex_hull_complex_gates(qcm::QuantumCircuitModel)
 
                     num_ex_pt = size(vertices_convex_hull)[1]
                     
-                    # Add convex hull facets if num_ex_pt == 2
+                    # Add the convex hull facet if num_ex_pt == 2
 
                     if num_ex_pt >= 3