+++ src/log.jl
@@ -39,7 +39,7 @@ function visualize_solution(results::Dict{String, Any}, data::Dict{String, Any};
         
         printstyled("  ","Total number of elementary gates (after presolve): ",size(data["gates_real"])[3],"\n"; color = :cyan)
         
-        printstyled("  ","Maximum depth of decomposition: ", data["depth"],"\n"; color = :cyan)
+        printstyled("  ","Maximum depth of decomposition: ", data["maximum_depth"],"\n"; color = :cyan)
         
         printstyled("  ","Input elementary gates: ", data["elementary_gates"],"\n"; color = :cyan)
 
@@ -52,6 +52,8 @@ function visualize_solution(results::Dict{String, Any}, data::Dict{String, Any};
         # printstyled("  ","Input target gate: ", data["target_gate"]["type"],"\n"; color = :cyan)
         
         printstyled("  ","Type of decomposition: ", data["decomposition_type"],"\n"; color = :cyan)
+
+        printstyled("  ","MIP optimizer: ", results["optimizer"],"\n"; color = :cyan)
 
         printstyled("\n","Optimal Circuit Decomposition","\n","\n"; color = :red)
         
@@ -137,7 +139,7 @@ function get_postprocessed_decomposition(results::Dict{String, Any}, data::Dict{
     gates_sol = Array{String,1}()
     id_sequence = Array{Int64,1}()
 
-    for d = 1:data["depth"]
+    for d = 1:data["maximum_depth"]
         id = findall(isone.(round.(abs.(results["solution"]["z_onoff_var"][:,d]), digits=3)))[1]
         push!(id_sequence, id)