+++ src/objective.jl
@@ -6,7 +6,7 @@ function objective_minimize_total_depth(qcm::QuantumCircuitModel)
     
     n_r          = size(qcm.data["gates_real"])[1]
     n_c          = size(qcm.data["gates_real"])[2]
-    depth        = qcm.data["depth"]
+    depth        = qcm.data["maximum_depth"]
     identity_idx = qcm.data["identity_idx"]
     num_gates    = size(qcm.data["gates_real"])[3]
 
@@ -53,7 +53,7 @@ function objective_minimize_cnot_gates(qcm::QuantumCircuitModel)
     
     n_r      = size(qcm.data["gates_real"])[1]
     n_c      = size(qcm.data["gates_real"])[2]
-    depth    = qcm.data["depth"] 
+    depth    = qcm.data["maximum_depth"] 
     cnot_idx = qcm.data["cnot_idx"]
 
     decomposition_type = qcm.data["decomposition_type"]